+++ Routes/Uploads/ProfilePic/Client_ProfilePic_Delete.js
@@ -15,11 +15,18 @@ const uploadClientProfilePic = async (req, res) => {
     try {
         const { userId } = req.body;
         const client = await Clients.findOne({ where: { id: userId } });
+        if (!client) {
+            return res.status(404).send({
+                message: "Client not found for the given userId",
+            });
+        }
         if (client.profile_pic_link) {
             const previousFilename = client.profile_pic_link.split("/").pop();
             const previousImagePath = `public/ProfilePics/${previousFilename}`;
             try {
-                fs.unlinkSync(previousImagePath);
+                if (fs.existsSync(previousImagePath)) {
+                    fs.unlinkSync(previousImagePath);
+                }
             } catch (error) {
                 return res.status(400).send({
                     message:
